A user can check out a book
- Removes original checked_out and checked_out_to database fields
in favor or a separate CheckOut object that is associated with
a book. This allows the app to keep a record of every time a book
was checked out, and possibly use the same CheckOut model for other
types of materials in the future.

- Add a Circulation controller class that interacts with the Book
and CheckOut models

- Add custom queries for the CheckOut class to return books
currently checked out

- Add virtual attributes to books to determine whether a book is
checked out and who it is checked out to

- Add validations to ensure a book is only checked out once