Add authorization plug and protect routes
This changes the test files to add a current user before testing
protected routes. When an unregistered user navigates to an unprotected
route, they will be redirected to the home page. Once they have been
logged in, they will be redirected back to the route they were trying to
access.